Rename login App component and fix import
