+++ app/src/client/login/components/index.ts
@@ -1,11 +1,11 @@
-import App from 'client/login/components/App';
 import AppleLinkBox from 'client/login/components/AppleLinkBox';
 import GitHubLinkBox from 'client/login/components/GitHubLinkBox';
 import GoogleLinkBox from 'client/login/components/GoogleLinkBox';
+import LoginApp from 'client/login/components/LoginApp';
 import MicrosoftLinkBox from 'client/login/components/MicrosoftLinkBox';
 
 export {
-    App, // force formatting
+    LoginApp, // force formatting
     AppleLinkBox,
     GitHubLinkBox,
     GoogleLinkBox,
